requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.30.1
pydantic==1.10.13
python-dotenv==1.0.1
neo4j==5.19.0
google-generativeai==0.7.2

